input validatio name with -
